Revert "feat(cli): posthog feature flags + fast_provision experiment (#3366)"
This reverts commit f7652de45b60b2c216aabf2de96dc05408e26ec5.

Rolling back on request. Net effect on users:

  * No more /decide fetch at startup — the up-to-1.5s cold-cache path
    is gone for everyone.
  * `fast_provision` experiment is off. Users who did NOT pass
    --beta/--fast get the old provisioning defaults (no tarball/images)
    instead of the test-variant defaults.
  * The persistent install-id file at ~/.config/spawn/.telemetry-id
    continues to exist — telemetry.ts is restored to its pre-#3366
    inline loadOrCreateUserId() that writes the same path, so existing
    users keep their PostHog identity.

Also reverts the follow-up review fixes in that PR (pick/feedback
fast-path skip and real SWR) since they only exist because the feature
flag system existed.

The feature-flags-cache.json left on disk by any user who ran the
shipped version becomes dead data; it's ignored when the module is
gone. No migration needed.
