+++ packages/cli/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@openrouter/spawn",
-  "version": "1.0.23",
+  "version": "1.0.24",
   "type": "module",
   "bin": {
     "spawn": "cli.js"
